Fix component status table
