+++ content/guides/status.mdx
@@ -3,6 +3,7 @@ title: Component status
 description: Check the status of accessibility reviews and the component lifecycle status of each component.
 ---
 
+import {RailsProvider} from '../../src/components/rails-provider'
 import {StatusTable} from '../../src/components/status'
 
 <Note>
@@ -11,4 +12,6 @@ Check out the [component lifecycle](/guides/component-lifecycle) for more inform
 
 </Note>
 
-<StatusTable />
+<RailsProvider>
+  <StatusTable />
+</RailsProvider>
