+++ package.json
@@ -41,6 +41,7 @@
     "yarn": "^1.22.18"
   },
   "devDependencies": {
+    "esm": "^3.2.25",
     "@github/markdownlint-github": "^0.3.0",
     "markdownlint-cli2": "^0.5.1",
     "markdownlint-cli2-formatter-pretty": "^0.0.3",
